+++ front/src/components/organisms/ResourceCardList.tsx
@@ -6,7 +6,7 @@ import CardResource from './CardResource'
 import { useSortByDate } from '../../hooks/useSortByDate'
 import { useSortByVotes } from '../../hooks/useSortByVotes'
 import { useAuth } from '../../context/AuthProvider'
-import { TResource, TFilters } from '../../types'
+import { TResource, TFilters, TSortOrder } from '../../types'
 import { useGetResources } from '../../hooks'
 
 const SpinnerStyled = styled(Spinner)`
@@ -27,11 +27,9 @@ const StyledText = styled(Text)`
   margin: 2rem;
 `
 
-type SortOrder = 'asc' | 'desc'
-
 type TResourceCardList = {
   filters: TFilters
-  sortOrder: SortOrder
+  sortOrder: TSortOrder
   handleAccessModal: () => void
   isSortByVotesActive: boolean
   onSelectedSortOrderChange: (selectedSortOrder: Array<TResource>) => void
@@ -70,12 +68,16 @@ const ResourceCardList: FC<TResourceCardList> = ({
             url={resource.url}
             description={resource.description}
             voteCount={resource.voteCount}
-            createdBy={resource.user.name}
+            createdBy={resource.user ? resource.user.name : ''}
             createdAt={resource.createdAt}
             updatedAt={resource.updatedAt}
             handleAccessModal={handleAccessModal}
             isFavorite={user ? resource.isFavorite : false}
-            editable={user?.name === resource.user.name}
+            editable={
+              resource.user
+                ? user?.name === resource.user.name
+                : user?.id === resource.userId
+            }
             resourceType={resource.resourceType}
             topics={resource.topics}
           />