+++ front/src/components/organisms/index.ts
@@ -6,7 +6,7 @@ import Login from './Login'
 import { MyFavoritesList } from './MyFavoritesList'
 import { MyResources } from './MyResources'
 import { Navbar } from './Navbar'
-import { ResourceCardList} from './ResourceCardList'
+import { ResourceCardList } from './ResourceCardList'
 import Register from './Register'
 import ResourceForm from './ResourceForm'
 import { SettingsManager } from './SettingsManager'
@@ -14,6 +14,7 @@ import { TopicsManagerBoard } from './TopicsManagerBoard'
 import { TopicsRadioWidget } from './TopicsRadioWidget'
 import type { TResourceForm } from './ResourceForm'
 import { UserAccessHome } from './UserAccessHome'
+import { UserProfileResourcesWidget } from './UserProfileResourcesWidget'
 import { VotesDateController } from './VotesDateController'
 
 export {
@@ -32,7 +33,8 @@ export {
   TopicsManagerBoard,
   TopicsRadioWidget,
   UserAccessHome,
+  UserProfileResourcesWidget,
   VotesDateController,
 }
 
-export type { TResourceForm}
+export type { TResourceForm }