+++ front/CHANGELOG.md
@@ -2,7 +2,30 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.3.1] - 2023-11-16
+
+## [0.3.1] - 2023-11-20
+
+### Added
+
+- Resources and favorites widgets to User Profile
+
+### Changed
+
+- User Profile UI
+- SortOrder type to avoid repetitions
+
+### Fixed
+
+- My Resources widget in Category page
+
+## [0.3.0] - 2023-11-16
+
+## Added
+
+- New resource search bar UI implementation
+
+## [0.2.1] - 2023-11-16
+
 
 ### Added